requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pymupdf
python-dotenv
streamlit
